+++ R/emmGrid-methods.R
@@ -1011,7 +1011,7 @@ regrid = function(object, transform = c("response", "mu", "unlink", "none", "pas
         pargs = object@grid[names(object@levels)]
         lbls = do.call(paste, c(pargs, sep = "."))
         if (!is.null(disp <- object@misc$display)) {  # fix up for the bookkeeping in nested models
-            object@V = object@V[disp, disp, drop = FALSE]
+            object@V = object@V[estble, estble, drop = FALSE]
             object@linfct = matrix(0, nrow = length(disp), ncol = length(estble))
             object@linfct[disp, ] = diag(1, length(estble))
             lbls = lbls[disp]