+++ inst/extexamples/qdrg-biglm.R
@@ -7,4 +7,4 @@ rg1 <- qdrg(log(conc) ~ source + factor(percent), data = pigs,
 emmeans(rg1, "source", type = "response")
 
 ## But in this particular case, we could have done it the easy way:
-##     rg1 <- qdrg(object = bigmod, data = pigs)
+##     rg1 <- qdrg(object = bigmod)