+++ inst/extexamples/qdrg-ordinal.R
@@ -3,4 +3,4 @@ wine.clm <- clm(rating ~ temp * contact, data = wine)
 ref_grid(wine.clm)
 
 # verify that we get the same thing via qdrg:
-qdrg(object = wine.clm, data = wine, ordinal.dim = 5)
+qdrg(object = wine.clm, ordinal.dim = 5)