+++ R/helpers.R
@@ -784,10 +784,13 @@ emm_basis.geese = function(object, trms, xlev, grid, vcov.method = "vbeta", ...)
 
 
 ### glmtoolbox package -- we just pretend it's lm and it seems to work
+
+#' @exportS3Method recover_data glmgee
 recover_data.glmgee = function(object, ...) {
     recover_data.lm(object, ...)
 }
 
+#' @exportS3Method emm_basis glmgee
 emm_basis.glmgee = function(object, trms, xlev, grid, vcov.method = "robust", ...) {
     vcov. = vcov(object, type = vcov.method, ...)
     class(object) = c("glm", "lm") # pretend it's glm